Got the project started
Today I was able to successfully run the python quickstart resource that was on google's calendar api starter kit.

I created the project on google developer's platform, I created the api token and auth process as well as added the credentials.json package to the working directory of the project.