First draft of correlation coefficient method
Through using the mathematical technique I proposed on the notion page I was able to code an algorithm that compare the template schedule with the user schedule to determine a correlation coefficient. In this push I've written the first draft to that correlation coefficient through rough observation I seem to get a better score when the schedules align and a worse score when the schedules deviate.