Debugging
No major updates to contribute within this update. I found that I was getting a negative relation coefficient so I went in adding in print statements to find out what could have been wrong. Through this process I've found out that there's nothing wrong with the code but there was a misunderstanding on my side.