Made schedule visualization
I actually still have that problem with the negative coefficient so to fix that I've resorted to visualizing the schedule using matplotlib.

In this push update I've managed to create a new python script file that graphs the google calendar schedule.

I will return and analyze the visualization to pinpoint where this error could be originating from.