Template can now handle when event is missing in the user calendar
In this commit I was able to get the code to work even when the event is missing in the user schedule by "regression to the mean". If the event isn't found a placeholder will be used as the "pair" for the template event and later when the mean time and duration is found the code will replace all of the missing (placeholder) events with the mean values making them appear on the line.