Minor change to invalid creds
I believe that I accidently indented one of the conditional statements pertaining to the invalid credentials conditional statement.
I have fixed the indent. The code now runs such that if you have an invalid token and you delete it, you'll be prompted to create a new one through the google authentification system.